% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlR.toolbox.R
\name{clearConsole}
\alias{clearConsole}
\title{Function clearConsole()}
\usage{
clearConsole()
}
\description{
This function clears the console.
}
\examples{
# clearConsole()
#   clears the console
}
\author{
Jean Marc Lietaer, \email{jmlietaer.tech@gmail.com}
}
\keyword{utilities}
