% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlR.toolbox.R
\name{packages.remove}
\alias{packages.remove}
\title{Function packages.remove()}
\usage{
packages.remove(package_name_list = "")
}
\arguments{
\item{package_name_list}{List of package(s) to remove. Default: <empty>}
}
\value{
A vector of character strings representing the result of the operation.
}
\description{
This function removes a list of packages
}
\examples{
# pckgs=c("alfa", "beta", "gamma")
# packages.remove(pckgs)
}
\author{
Jean Marc Lietaer, \email{jmlietaer.tech@gmail.com}
}
\keyword{utilities}
