% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlR.toolbox.R
\name{packages.unload}
\alias{packages.unload}
\title{Function packages.unload()}
\usage{
packages.unload(package_name_list = "")
}
\arguments{
\item{package_name_list}{List of package(s) to unload. Default: <empty>}
}
\value{
Returns a vector of character strings representing the result of the operation.
}
\description{
This function unloads a list of packages
}
\examples{
# pckgs=c("alfa", "beta", "gamma")
# packages.unload(pckgs)
}
\author{
Jean Marc Lietaer, \email{jmlietaer.tech@gmail.com}
}
\keyword{utilities}
