% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlR.toolbox.R
\name{system.info}
\alias{system.info}
\title{Function system.info()}
\usage{
system.info()
}
\value{
Returns: [R version] [R platform] [system date and time] [system timezone]
}
\description{
This function returns the following information:
}
\examples{
system.info()
}
\author{
Jean Marc Lietaer, \email{jmlietaer.tech@gmail.com}
}
\keyword{utilities}
