% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmlR.toolbox.R
\name{write.graphml}
\alias{write.graphml}
\title{Function write.graphml()}
\usage{
write.graphml(orig, rel = "N", directed = TRUE, sep = ";", data_in,
  header = T, fill_color = T, data_out = "diagram.graphml",
  applic = "no")
}
\arguments{
\item{orig}{Origin of the input, "F"=file, "D"=data. Default: <empty>}

\item{rel}{Type of relation. "N"=numeric, "C"=character. Default: "N"}

\item{directed}{TRUE or FALSE - Directed graph. Default: "TRUE"}

\item{sep}{Separator character. Default: ";"}

\item{data_in}{<name_of_input_variable> | <fully_qualified_file_name_of_input_file>}

\item{header}{TRUE or FALSE - Input had a header. Default: "TRUE"}

\item{fill_color}{TRUE or FALSE - Input contains fill color (format #FFFFFF). Default: "TRUE"}

\item{data_out}{<fully_qualified_file_name_of_output_file>}

\item{applic}{<fully_qualified_file_name_of_application_to_open_data_out>}
}
\value{
Returns a vector of character strings representing the result of the operation.
}
\description{
This function produces a GraphML file based on user input (file or data-variable).
}
\examples{
# rc <- jmlR.toolbox::write.graphml(orig = "F", rel = "C", sep = ";", directed=FALSE, data_in = "input.csv",
# header = TRUE, fill_color = F, data_out =  paste(getwd(), "/", "output.graphml", sep = ""),
# applic = "/home/jmlietaer/yEd/yEd")
}
\author{
Jean Marc Lietaer, \email{jmlietaer.tech@gmail.com}
}
\keyword{analysis}
\keyword{graph,}
\keyword{network}
\keyword{social}
\keyword{utilities,}
